requirements-checky.txt
# Checky Multimodal Assistant Requirements
# Core framework
pipecat-ai[google,websocket]>=0.0.80

# Web framework and API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Rate limiting and security
slowapi>=0.1.5
bcrypt>=4.0.1

# Database and utilities
loguru>=0.7.0
python-multipart>=0.0.6

# Optional development dependencies
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# httpx>=0.25.0
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0